Add files via upload
In affine cipher each letter in an alphabet is mapped to its numeric equivalent x, encrypted using a simple mathematical function, and converted back to a letter. Letter A is given number 0 and letter Z is given number 25. Each letter is encrypted with the function (5x + 8) mod 26. The decryption function is 21(y − 8) mod 26.